fixed interger narrowing bug in flowedge

+++ graphs/flowedge.cc
@@ -13,7 +13,7 @@ struct edge {
 };
 vector<edge> edges;
 vvi adj;
-void addEdge(int from, int to, int c, int f = 0) {
+void addEdge(int from, int to, ll c, ll f = 0) {
   edges.eb(from, to, c, f);
   adj[from].pb(SZ(edges) - 1);
   adj[to].pb(SZ(edges) - 1);
